UNDO FUNCTION

+++ project/project/main.cpp
@@ -3,6 +3,7 @@
 #include <fstream>
 #include <vector>
 #include <map>
+#include <stack>
 
 #include "player.h"
 #include "game.h"
@@ -22,7 +23,7 @@ string findPlayer(string name)
 		while (getline(ifile, line))
 		{
 			temp = "";
-			for (int i = 0; i < line.size(); i++)
+			for (size_t i = 0; i < line.size(); i++)
 			{
 				if (line.at(i) == ' ')
 					i = line.size();
